Fixing deprecation warning/issue in Rails 6.1 with multiple formats, removing Rails < 4 support
